requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
bidict==0.23.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.38.2
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
yarl==1.18.3
